Prepare for release 2011.11.11
This release it dedicated to the brave pandora engineers who fight for
love, peace and transport layer encrypted plaintext password logins. To
achieve their goals they climbed the highest mountain, dived the deepest
ocean and explained Diffie-Hellman to their management. But they also
had to make terrible sacrifices to succeed: pianobar, a tiny but useful
tool did not survive the fight without serious injuries. `git log` will
never forget your efforts to create a brighter future. For all of us.

Expect this release to be not as stable as previous releases as it
contains experimental code that was not meant to be released at this
point.