Update ui_act.c
I accidentally hit the '-' button on my keyboard and the pianobar program just happened to be in the foreground. It banned by song, and I couldn't figure out an easy way in the Pandora UI to un-ban the song....and it was a song that I had favorited, so I couldn't stand to not be able to listen to it for a month. I had to delete and re-create the station. I put in this confirm message to make sure I mean to ban a song before actually doing it.