Switch back to libcurl
Drops libwaitress. Adds the new dependency libcurl and drops gnutls.

I wouldn’t say writing my own HTTP library was a mistake – it was not
and the experience gained was worth it. Instead I have to acknowledge
that libcurl is just better than my own implementation. Sure, it does a
lot more than HTTP – one could call that bloat. Yet if you just want to
get the job done™ reusing code is the way to go.

See #512 and #513.