Add login retry functionality
This patch adds a new setting in the config file: num_login_tries.  The
default value is 1 to match the current functionality.  If the value is
set larger, the program will retry up to this many times until there is
a successful login, or the number of tries has been exhausted.

In order to correctly accomplish this new piece of functionality, I had
to add two values to the settings struct to keep track of the values of
the username and password from the config file separate from those that
have been created by looking them up so that they can be reset upon
failure.

Lightly tested without any problems.